# Slightly improve perf of tan & tanh, and accuracy too