@rgommers 2015-03-08
@calixteman I'll close this PR now due to no activity. Please reopen and reply to @seberg's request if you think this PR is still valid.